style: create RequestItemService interface
